+++ MicrocontrollerSystem/Release/box/State.d
@@ -1,0 +1,3 @@
+box/State.d box/State.o: ../box/State.cpp ../box/State.h
+
+../box/State.h:
