+++ MicrocontrollerSystem/Release/makedep.mk
@@ -34,5 +34,5 @@ interface\Istream.cpp
 
 main.cpp
 
-remote\State.cpp
+box\State.cpp
