+++ MicrocontrollerSystem/main.cpp
@@ -1,10 +1,10 @@
 #include "atmicro.h"
-#include "remote/parse.h"
+#include "box/parse.h"
 int main(void)
 {
-	Keypad pad(IOPORTA);
-	Istream inp;
 	ParallelTextLCD lcd(IOPORTB, IOPIND2, IOPIND7, IOPIND5);
+	Keypad pad(IOPORTC);
+	Istream inp;
 	while (1)
 	{
 		inp.update(pad);
